requirements.txt
click==7.1.2
fido2==0.9.1
pyserial==3.5
